add tooltip for checkbox, checkbox removed from header
