+++ src/ducks/Watchlists/Widgets/Table/AssetsToggleColumns.module.scss
@@ -22,6 +22,12 @@
   }
 }
 
+.columnId {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+}
+
 .assetCheckbox {
   margin: auto;
 }
